+++ MoneyBuddi/src/model/Category.java
@@ -41,8 +41,8 @@ public class Category {
 		return id;
 	}
 	
-	public void setId(int id) {
-		this.id = id;
+	public void setId(long id) {
+		this.id = (int)id;
 	}
 	
 	public String getCategory() {
